+++ src/components/ui/NavButton/NavButton.tsx
@@ -1,22 +1,12 @@
 import clsx from 'clsx';
 import Link from 'next/link';
-import type { FC, ButtonHTMLAttributes } from 'react';
+import type { FC } from 'react';
 
 import styles from './NavButton.module.css';
-
-interface BaseProps {
-  isActive?: boolean;
-  className?: string;
-  children: React.ReactNode;
-}
-
-type ButtonProps = BaseProps & ButtonHTMLAttributes<HTMLButtonElement> & { href?: undefined };
-type LinkProps = BaseProps & { href: string; onClick?: () => void };
-
-type NavButtonProps = ButtonProps | LinkProps;
+import type { NavButtonProps, ButtonProps } from './types';
 
 const NavButton: FC<NavButtonProps> = ({ isActive = false, className, children, ...props }) => {
-  const classes = clsx(styles['nav-button'], isActive && styles.active, className);
+  const classes = clsx(styles['navButton'], isActive && styles.active, className);
 
   if ('href' in props && props.href) {
     return (
